requirements.txt
# # Core LangChain and LangGraph dependencies
# langchain==0.1.16
# langchain-openai==0.1.1
# langgraph==0.0.40
# langchain-core==0.1.42

# # Azure OpenAI
# openai==1.14.0

# # Utilities
# python-dotenv==1.0.1
# typing-extensions==4.10.0

# # Tools (optional, for extended examples)
# tavily-python==0.3.0  # Web search
# requests==2.31.0


# Core LangChain and LangGraph dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-core>=0.1.0

# Azure OpenAI
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# Tools (optional, for extended examples)
tavily-python>=0.2.0  # Web search (optional)
requests>=2.28.0
